[UPDATE]: error validations in useCases

+++ src/useCases/comments.use.js
@@ -1,3 +1,4 @@
+import { StatusHttp } from "../libs/statusHttp.js";
 import { Comment } from "../models/comment.model.js";
 import { Post } from "../models/posts.model.js";
 
@@ -8,22 +9,23 @@ async function getAll() {
 // POST /cooments/
 async function addComment(newComment, userCurrent) {
   console.log({ ...newComment, author: userCurrent });
-
   let commentCreated = await Comment.create({...newComment,author: userCurrent,
   });
-
   await Post.findByIdAndUpdate(newComment.post_id, {
     $push: { comments: commentCreated._id },
   });
-
   return commentCreated;
 }
 
-function deleteById(idComment) {
+async function deleteById(idComment) {
+  const commentFinded = await Comment.findById(idComment)
+  if(!commentFinded) throw new StatusHttp('Comment not found', 400)
   return Comment.findByIdAndDelete(idComment);
 }
 
-function update(idComment, unupdatedComment) {
+async function update(idComment, unupdatedComment) {
+  const commentFinded = await Comment.findById(idComment)
+  if(!commentFinded) throw new StatusHttp('Comment not found', 400)
     return Comment.findByIdAndUpdate(idComment, unupdatedComment, { new: true })
 }
